+++ packages/core-data/src/reducer.js
@@ -238,6 +238,28 @@ export function userPermissions( state = {}, action ) {
 	return state;
 }
 
+/**
+ * Reducer returning autosaves keyed by their parent's post id.
+ *
+ * @param  {Object} state  Current state.
+ * @param  {Object} action Dispatched action.
+ *
+ * @return {Object} Updated state.
+ */
+export function autosaves( state = {}, action ) {
+	switch ( action.type ) {
+		case 'RECEIVE_AUTOSAVE':
+			const { postId, autosave } = action;
+
+			return {
+				[ postId ]: autosave,
+				...state,
+			};
+	}
+
+	return state;
+}
+
 export default combineReducers( {
 	terms,
 	users,
@@ -246,4 +268,5 @@ export default combineReducers( {
 	entities,
 	embedPreviews,
 	userPermissions,
+	autosaves,
 } );